requirements.txt
discord.py
aiohttp
pytz
python-dotenv
